Move requirement task list view to app.main.views.requirement_task_list
